Imprecise

Documentation

class (Num (Estimate a), Ord (Estimate a), Imprecise (Estimate a), Estimate (Estimate a) ~ Estimate a) => Imprecise a where

A class for finite-precision types.

Minimal complete definition

limiting, scale

Associated Types

type Estimate a

An estimate of values of type a suitable for estimating its size and precision.

Methods

precision :: a -> Estimate a

Estimate the precision of a value. precision a > precision b indicates that a is less precise than b. The default implementation is suitable unless Estimate a = a.

limiting :: a -> Estimate a

Estimate the size of the precision-limiting element of a collection of values. limiting must obey

precision a == precision (limiting a)

For floating-point numbers in the usual representations, limiting = abs. For collections of imprecise values, this is generally the greatest limiting value of all the elements.

scale :: Estimate a -> a -> a

Scale a value of by an estimate. scale must obey

limiting (scale x a) = abs x * limiting a

because most error-aware algorithms assume linearity of precision.